Introduce IEventReceiverInterface
In order to be able to AddEventReceiver without having to specify the event (redundant), created interface and abstract base class for EventReceiver.

Chose not to use a Flag Interface (which would have removed the need to create abstract base ) in favor of tidying up the API and also providing some validation if the implicit interface is used.